+++ example-app/daemons/process-long.js
@@ -1,14 +1,14 @@
+`user strict`;
 (async function() {
+const path = require('path')
 
 console.log('Long running process sample', __filename)
 console.log('this can be a Kafka producer... listen to incoming tcp messages and pushing to topic')
 console.log('Do take note limitations for Long running NodeJS process')
 console.log('Can also be for cronjobs (but better to use cronjob call an API)')
 
-await require('@es-labs/node/config')(process.cwd()) //  first thing to run
+await require('@es-labs/node/config')(path.join(process.cwd(),'..')) //  first thing to run
 
-// mixing ES Modules into a CommonJS project
-// https://codewithhugo.com/use-es-modules-in-node-without-babel/webpack-using-esm/
 const { sleep } = require('esm')(module)('@es-labs/esm/sleep')
 
 const run = async () => {
@@ -22,29 +22,5 @@ const run = async () => {
 }
 
 run().catch(e => console.error(`[***] ${e.message}`, e))
-const errorTypes = ['unhandledRejection', 'uncaughtException']
-const signalTraps = ['SIGTERM', 'SIGINT', 'SIGUSR2']
-
-errorTypes.forEach(type =>
-  process.on(type, async e => {
-    try {
-      console.log(`process.on ${type}`)
-      console.error(e)
-      process.exit(0)
-    } catch (_) {
-      process.exit(1)
-    }
-  })
-)
-
-signalTraps.forEach(type =>
-  process.once(type, async () => {
-    try {
-      await consumer.disconnect()
-    } finally {
-      process.kill(process.pid, type)
-    }
-  })
-)
-
+require('./traps')(async (type) => console.log(`Signal ${type}`))
 }())