+++ example-app/sandbox/tcpServer.js
@@ -41,40 +41,11 @@ server.on('close',function(){
   console.log('Server closed 2')
 })
 
-const errorTypes = ['unhandledRejection', 'uncaughtException']
-const signalTraps = ['SIGTERM', 'SIGINT', 'SIGUSR2']
+// for (let socket of sockets) socket.destroy()
+// server.close(function () { server.unref() })
 
-errorTypes.forEach(type => {
-  process.on(type, async e => {
-    try {
-      console.log(`process.on ${type}`)
-      console.error(e)
-      return process.exit(0)
-    } catch (_) {
-      return process.exit(1)
-    }
-  })
-})
 
-signalTraps.forEach(type => {
-  process.once(type, async () => {
-    try {
-      console.log('Signal Trap', type)
-      for (let socket of sockets) socket.destroy()
-      // server.close()
-      server.close(function () {
-        console.log('server closed')
-        server.unref()
-      })
-      // await consumer.disconnect()
-      console.log('end')
-    } finally {
-      // Do not call this as need time to close the server...
-      // process.kill(process.pid, type)
-    }
-    return
-  })
-})
+
 
 /*
 // Import net module.